Abstract some logic
