+++ src/pages/steps/CultuurkuurLabelsPicker.tsx
@@ -12,14 +12,14 @@ type Props = {
   data: HierarchicalData[];
   selected?: string[];
   onConfirm?: (newSelected: string[]) => void;
-  translationKey?: string;
+  translationKey: string;
 };
 
 const CultuurkuurLabelsPicker = ({
   data,
   selected = [],
   onConfirm = () => ({}),
-  translationKey = 'location',
+  translationKey,
 }: Props) => {
   const { t } = useTranslation();
   const [
